pyproject.toml
[project]
name = "market-data-pipeline"
version = "0.1.0"
description = "주식의 과거 데이터 수집, 그리고 일간 데이터 배치 파이프라인"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "polars>=0.20.31",
    "finance-datareader>=0.9.31",
    "psycopg2-binary>=2.9.9",
    "sqlalchemy>=2.0.25",
    "python-dotenv>=1.0.0",
    "schedule>=1.2.0",
    "requests>=2.31.0",
]

[project.optional-dependencies]
dev = [
    "pytest==7.4.2",
    "black==23.7.0",
    "flake8==6.0.0",
    "mypy==1.5.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.black]
line-length = 88
target-version = ['py39']

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true